module9-task2: refactor
