+++ src/store/action.ts
@@ -2,5 +2,3 @@ import { createAction } from '@reduxjs/toolkit';
 import {AppRoute} from '../enums/app-route.ts';
 
 export const redirectToRoute = createAction<AppRoute | string>('redirectToRoute');
-
-
