+++ src/utils/format-time.ts
@@ -16,10 +16,10 @@ export const formatRunTime = (runTime: number) => {
 export function calcRemainingTime(duration: number, currentTime: number) {
   const time = duration - currentTime;
   const hours = Math.floor(time / SEC_PER_HOUR);
-  const minuts = Math.floor((time % SEC_PER_HOUR) / SEC_PER_MINUTE);
-  const seconds = Math.floor(time - hours * SEC_PER_HOUR - minuts * SEC_PER_MINUTE);
+  const minutes = Math.floor((time % SEC_PER_HOUR) / SEC_PER_MINUTE);
+  const seconds = Math.floor(time - hours * SEC_PER_HOUR - minutes * SEC_PER_MINUTE);
 
   return hours > 0
-    ? `-${hours.toString().padStart(2, '0')}:${minuts.toString().padStart(2, '0')}:${seconds.toString().padStart(2, '0')}`
-    : `-${minuts.toString().padStart(2, '0')}:${seconds.toString().padStart(2, '0')}`;
+    ? `-${hours.toString().padStart(2, '0')}:${minutes.toString().padStart(2, '0')}:${seconds.toString().padStart(2, '0')}`
+    : `-${minutes.toString().padStart(2, '0')}:${seconds.toString().padStart(2, '0')}`;
 }
